refactor: decouple QIS trace replay from surface code
