+++ python/quantum-pecos/tests/pecos/test_hosted_operations.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import pytest
-from pecos.qec.surface.decode import _validate_trace_hosted_operations_if_requested
+from pecos._qis_trace_replay import _validate_trace_hosted_operations_if_requested
 from pecos.quantum.hosted import validate_hosted_operations
